requirements.txt
asgiref==3.11.0
boto3==1.41.5
botocore==1.41.5
dj-database-url==3.1.2
Django==5.0.3
django-storages==1.14.6
jmespath==1.0.1
pillow==12.0.0
psycopg[binary]==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0